requirements.txt
streamlit>=1.30.0
google-generativeai
openai
selenium
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
urllib3<2.0.0
webdriver-manager
